+++ src/ctap2.rs
@@ -11,6 +11,7 @@ use crate::{sizes::*, Bytes, TryFromStrError};
 
 pub use crate::operation::{Operation, VendorOperation};
 
+pub mod authenticator_config;
 pub mod client_pin;
 pub mod credential_management;
 pub mod get_assertion;
@@ -45,6 +46,8 @@ pub enum Request<'a> {
     Selection,
     // 0xC
     LargeBlobs(large_blobs::Request<'a>),
+    // 0xD
+    AuthenticatorConfig(authenticator_config::Request<'a>),
     // vendor, to be embellished
     // Q: how to handle the associated CBOR structures
     Vendor(crate::operation::VendorOperation),
@@ -118,10 +121,14 @@ impl<'a> Request<'a> {
                 Request::LargeBlobs(cbor_deserialize(data).map_err(CtapMappingError::ParsingError)?)
             }
 
+            Operation::Config => Request::AuthenticatorConfig(
+                cbor_deserialize(data).map_err(CtapMappingError::ParsingError)?,
+            ),
+
             // NB: FIDO Alliance "stole" 0x40 and 0x41, so these are not available
             Operation::Vendor(vendor_operation) => Request::Vendor(vendor_operation),
 
-            Operation::BioEnrollment | Operation::PreviewBioEnrollment | Operation::Config => {
+            Operation::BioEnrollment | Operation::PreviewBioEnrollment => {
                 debug_now!("unhandled CBOR operation {:?}", operation);
                 return Err(CtapMappingError::InvalidCommand(op).into());
             }
@@ -143,6 +150,7 @@ pub enum Response {
     Selection,
     CredentialManagement(credential_management::Response),
     LargeBlobs(large_blobs::Response),
+    AuthenticatorConfig,
     // Q: how to handle the associated CBOR structures
     Vendor,
 }
@@ -161,7 +169,7 @@ impl Response {
             GetAssertion(response) | GetNextAssertion(response) => cbor_serialize(response, data),
             CredentialManagement(response) => cbor_serialize(response, data),
             LargeBlobs(response) => cbor_serialize(response, data),
-            Reset | Selection | Vendor => Ok([].as_slice()),
+            Reset | Selection | AuthenticatorConfig | Vendor => Ok([].as_slice()),
         };
         if let Ok(slice) = outcome {
             *status = 0;
@@ -445,6 +453,11 @@ pub trait Authenticator {
         Err(Error::InvalidCommand)
     }
 
+    fn authenticator_config(&mut self, request: &authenticator_config::Request) -> Result<()> {
+        let _ = request;
+        Err(Error::InvalidCommand)
+    }
+
     /// Dispatches the enum of possible requests into the appropriate trait method.
     #[inline(never)]
     fn call_ctap2(&mut self, request: &Request) -> Result<Response> {
@@ -531,6 +544,15 @@ pub trait Authenticator {
                         debug!("error: {:?}", _e);
                     })?,
                 ))
+            }
+
+            // 0xD
+            Request::AuthenticatorConfig(request) => {
+                debug_now!("CTAP2.CFG");
+                self.authenticator_config(request).inspect_err(|_e| {
+                    debug!("error: {:?}", _e);
+                })?;
+                Ok(Response::AuthenticatorConfig)
             }
 
             // Not stable